use view hierachy for screenshots
