+++ flutter/lib/src/replay/scheduled_recorder.dart
@@ -2,10 +2,11 @@ import 'dart:async';
 import 'dart:ui';
 
 import 'package:meta/meta.dart';
+import 'package:sentry_flutter/src/replay/scheduled_recorder_config.dart';
 
 import '../../sentry_flutter.dart';
-import 'recorder.dart';
-import 'recorder_config.dart';
+import '../screenshot/recorder.dart';
+import '../screenshot/recorder_config.dart';
 import 'scheduler.dart';
 
 @internal
