+++ flutter/lib/src/replay/scheduled_recorder_config.dart
@@ -1,0 +1,11 @@
+import '../screenshot/recorder_config.dart';
+
+class ScheduledScreenshotRecorderConfig extends ScreenshotRecorderConfig {
+  final int frameRate;
+
+  const ScheduledScreenshotRecorderConfig({
+    super.width,
+    super.height,
+    required this.frameRate,
+  });
+}
