+++ flutter/test/replay/masking_config_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/replay/masking_config.dart';
+import 'package:sentry_flutter/src/screenshot/masking_config.dart';
 
 import 'test_widget.dart';
