+++ flutter/test/replay/widget_filter_test.dart
@@ -2,7 +2,7 @@ import 'package:flutter/services.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
-import 'package:sentry_flutter/src/replay/widget_filter.dart';
+import 'package:sentry_flutter/src/screenshot/widget_filter.dart';
 
 import 'test_widget.dart';
